# doc/openapi.yaml
openapi: 3.0.3
info:
  title: BrandingPo - API de Monitoreo de Medios, Clipping & Social Listening
  description: |
    Especificación OpenAPI de los endpoints de gestión de medios de prensa, autodescubrimiento, 
    clipping editorial y cuantificación de reacciones en Facebook (Fast-Flow) de BrandingPo.
  version: 1.1.0
servers:
  - url: http://brandingpo.test
    description: Servidor de Desarrollo Local
paths:
...
  /medios/{id}:
    put:
      summary: Actualizar medio de prensa existente
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MedioInput'
      responses:
        '302':
          description: Redirección con medio actualizado

    delete:
      summary: Eliminar medio y sus notas asociadas
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '302':
          description: Redirección tras eliminación
...
components:
  schemas:
    MedioInput:
      type: object
      required:
        - nombre
        - tipo_medio
      properties:
        nombre:
          type: string
          example: Diario de Cuyo
        tipo_medio:
          type: string
          enum: [digital, impreso, radio, tv]
          example: digital
        url_sitio:
          type: string
          format: uri
          nullable: true
          example: https://www.diariodecuyo.com.ar
        url_facebook:
          type: string
          format: uri
          nullable: true
          example: https://www.facebook.com/diariodecuyoweb
        avatar_url:
          type: string
          nullable: true
        feed_rss_url:
          type: string
          nullable: true
        alcance_tipo:
          type: string
          enum: [local, provincial, nacional]
          default: provincial
        sesgo_editorial_estimado:
          type: string
          enum: [oficial, opositor, independiente, neutral]
          default: neutral
...